📋 Restructured Commands 📋
Completely restructured the way commands are stored. Each command is now stored in a separate file in the 'commands' folder. I opted for this so that the main index.js feels less cluttered and if a command stops working it should cause less harm to the program as a whole.